% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjust_p.R
\name{Adjust_p}
\alias{Adjust_p}
\title{Adjusting p-values, auxiliary function}
\usage{
Adjust_p(ctp.struc, ctp.pval)
}
\arguments{
\item{ctp.struc}{Object of class \code{ctp.str}.}

\item{ctp.pval}{Data frame with hypothesis names and unadjusted p-values. Output from \code{CTPcompare}.}
}
\value{
Adjusted p-values
}
\description{
Adjusting p-values, auxiliary function
}
