% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjust_raw.R
\name{Adjust_raw}
\alias{Adjust_raw}
\title{Adjusting raw p-values of a CTP}
\usage{
Adjust_raw(
  ctp.struc,
  p.value,
  dataset.name = NULL,
  factor.name = NULL,
  factor.levels = NULL,
  model = NULL,
  test.name = NULL
)
}
\arguments{
\item{\code{ctp.struc}}{Object generated by \code{\link{IntersectHypotheses}}}

\item{\code{p.value}}{Vector of raw p-values in the order of the hypotheses created by \code{\link{summary.ctp.str}}}

\item{\code{dataset.name}}{Character string naming the analysis dataset (optional - only for documentation purposes).}

\item{\code{factor.name}}{Character string naming the factor whose levels are compared (optional - only for documentation purposes).}

\item{\code{factor.levels}}{Vector of type "character" containing the levels of the treatment factor
(optional - only for documentation purposes).}

\item{\code{model}}{Model used in the analysis (optional - only for documentation purposes).}

\item{\code{test.name}}{Character string naming the statistical test applied.}
}
\value{
An object of \code{oldClass = "ctp"} to be used for summarizing and plotting the results.
}
\description{
Function that adjusts the raw p-values of a closed testing procedure. The raw p-values are adjusted according to the closure principle.
The adjusted p-value is calculated as the maximum of the raw p-value from the current hypothesis in question and the raw p-values from
all subsequent hypotheses that contain the current hypothesis.
}
\seealso{
\code{\link{IntersectHypotheses}}, \code{\link{AnalyseCTP}}, \code{\link{Display}},
\code{\link{summary.ctp}}
}
