% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorectal}
\alias{colorectal}
\title{Data from a Study in Colorectal Cancer}
\format{
A data frame with 104 observations and 2 variables:
	\describe{
		 \item{\code{dose}}{a factor with levels \code{control} \code{low} \code{high}}
		 \item{\code{responder}}{a factor with levels \code{no} \code{yes}}
	}
}
\usage{
data(colorectal)
}
\description{
A dataframe containing the responders to the treatment with two doses of
an experimental drug or a standard treatment.
}
\keyword{datasets}
