% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctp_cbf.R
\name{ctp_cbf}
\alias{ctp_cbf}
\title{Graph Cerebral Blood Flow}
\usage{
ctp_cbf(folder, method = c("maxslope", "sSVD", "bSVD", "oSVD"),
  aif = c("auto", "max", "gv", "user"), t0, C0 = 1, a = 3, b = 1.5,
  aif.user, lag)
}
\arguments{
\item{folder}{directory containing the set of CT images to be combined. Must be arranged in alphanumerical order according to the order in which they were collected}

\item{method}{character string indicating estimation technique to be used ("maxslope","sSVD","bSVD", or "oSVD")}

\item{aif}{character string indicating arterial input function to be used ("auto", "max", "gv", or "user")}

\item{t0}{parameter for aif="gv". The time in which contrast enters. No default value. Required if aif="gv".}

\item{C0}{parameter for aif="gv". Constant for gv function. Defaults to 1.}

\item{a}{parameter for aif="gv". Shape parameter for gv function. Defaults to 3.}

\item{b}{parameter for aif="gv". Rate parameter for gv function. Defaults to 1.5.}

\item{aif.user}{user-defined vector for arterial input function. Required if aif="user".}

\item{lag}{time interval in seconds for collection of CT data. Usually set to 1 or 2. Required.}
}
\description{
This function generates a perfusion map of cerebral blood flow.
It should be used for research and illustrative purposes only and should NOT be used to diagnose/treat any disease.
}
\examples{
ctp.cbf(folder="C:/.../r001/loc12", method="maxslope", aif="gv", t0=10, lag=1) 

ctp.cbf(folder="C:/Users/slirette2/Desktop/r083/loc3", method="oSVD", aif="gv", t0=10, C0=1.13, lag=1)

svec = c(50,50,50,50,50,50,50,50,50,50,50,51.,211,365,445,446,396,323,247,181,127,87,58,50,50,50,50)
ctp.cbf(folder="C:/Users/slirette2/Desktop/r083/loc3", method="oSVD", aif="user", aif.user=svec, lag=1)

}
\keyword{blood}
\keyword{cbf}
\keyword{ctp}
\keyword{flow}
\keyword{perfusion}

