% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTPcompare.R
\name{CTPcompare}
\alias{CTPcompare}
\title{Comparisons of treatment levels}
\usage{
CTPcompare(CTPparms, ...)
}
\arguments{
\item{CTPparms}{Object of class \code{CTPparms} (i.e. generated by the function \code{getCTPparms}}

\item{...}{Additional arguments for the chosen test}
}
\value{
A data.frame with the hypotheses and the unadjusted p-values
}
\description{
Performes all required comparisons of treatment levels
}
