% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCTPparms.R
\name{getCTPparms}
\alias{getCTPparms}
\title{Auxiliary function}
\usage{
getCTPparms(
  ctp.struc,
  model,
  dataset,
  factor.name = NULL,
  test.name = "F",
  ...
)
}
\arguments{
\item{ctp.struc}{Object of class \code{ctp.str}.}

\item{model}{Formula of the form response~treatment.}

\item{dataset}{Dataframe.}

\item{factor.name}{Character string naming the factor whose levels are compared (treatment factor).
By default the first variable of the right-hand side of the model formula is used.}

\item{test.name}{One of the following strings \itemize{
    \item \code{"F"} - F-Test (ANOVA, default)
    \item \code{"kruskal"} -Kruskal-Wallis-Test
    \item \code{"chisq"} - Chi square test
    \item \code{"prob"} - Fisher's exact test for total number of observations <200 else Chi square test
    \item \code{"lgrank"} - Logrank-test
    \item \code{"jonckheere"} - Jonckheere-Terpstra test of ordered alternatives
    \item \code{"glm"} - generlaized linear model
}}

\item{...}{Additional arguments for the chosen test}
}
\value{
A list with CTP parameters
}
\description{
Extracting CTP - parameters
}
