% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponse.R
\name{getResponse}
\alias{getResponse}
\title{Auxiliary function}
\usage{
getResponse(formula, data)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Dataframe}
}
\value{
Vector of response
}
\description{
Getting the response vector out of an object of class 'formula'
}
\examples{

data(pasi)

data(pasi)
getResponse(pasi.ch ~ dose, data=pasi)

library(survival)
data(ovarian)
ovarian$subgroups <- as.factor(10*ovarian$ecog.ps+ovarian$rx)

surv <- getResponse(Surv(futime,fustat)~rx, data=ovarian)
print(surv)
class(surv)
}
