% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glucose}
\alias{glucose}
\title{Data from a Study in Diabetes II patients}
\format{
A data frame with 89 observations and 3 variables:
\describe{
	\item{\code{GLUCOSE.BLA}}{Fasting plasma glucose at baseline}
	\item{\code{GLUCOSE.CHANGE}}{Changes in fasting plasma glucose from baseline}
	\item{\code{DOSE}}{An ordered factor with levels \code{PLACEBO} < \code{LOW} < \code{MEDIUM} < \code{HIGH}}
}
}
\usage{
data(glucose)
}
\description{
The dataframe \code{glucose} comprises the baseline values of fasting plasma glucose (mmol/L)
and their changes from baseline to the end of the study.
}
\keyword{datasets}
