% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pasi}
\alias{pasi}
\title{Data from a Psoriasis study}
\format{
A data frame with 72 observations and 3 variables:
\describe{
    \item{treatment}{a factor with levels \code{Etretin10mg} \code{Etretin25mg} \code{Etretin50mg} \code{Placebo}}
    \item{pasi.ch}{Changes in PASI score within two months}
    \item{dose}{Dose of Etretin as ordered factor with levels
          \code{Placebo} < \code{ET.10mg} < \code{ET.25mg} < \code{ET.50mg}}
    }
}
\usage{
data(pasi)
}
\description{
The dataframe \code{pasi} comprises the changes in PASI-score (Psoirasis Area and Severity Index)
from Baseline within two months in 72 patients treated with different
doses of Etretin or Placebo in a double blind study.
}
\keyword{datasets}
