% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingSet.R
\name{TestingSet}
\alias{TestingSet}
\title{Testing set for an elementary hypothesis}
\usage{
TestingSet(ctp.struc, Hyp)
}
\arguments{
\item{\code{ctp.struc}}{Object of class \code{ctp.str}.}

\item{\code{Hyp}}{Elementary hypothesis (character variable).}
}
\value{
The testing set for the elementary hypothesis (character vector).
}
\description{
Testing set for an elementary hypothesis
}
\examples{

	three.to.first <- IntersectHypotheses(list(1:2,c(1,3),c(1,4)))
   Set13  <- TestingSet(three.to.first,Hyp="[13]")
	Set=13

}
