% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingSet.R
\name{TestingSet}
\alias{TestingSet}
\title{Testing set for an elementary hypothesis}
\usage{
TestingSet(ctp.struc, Hyp)
}
\arguments{
\item{ctp.struc}{Object of class \code{ctp.str}.}

\item{Hyp}{Elementary hypothesis (character variable).}
}
\value{
The testing set for the elementary hypothesis (character vector).
}
\description{
Testing set for an elementary hypothesis
}
\examples{

	Pairwise <- IntersectHypotheses(list(c(1,2), c(1,3),
	                                c(1,4), c(2,3), c(2,4), c(3,4)))
   Set24    <- TestingSet(Pairwise,"[24]")
   Set24
   
}
